refactor(devops): Remove openzeppelin dependency and folder and update dxswap-core with latest commi
