--- .openzeppelin/kovan.json
@@ -1,48 +1,0 @@
-{
-  "contracts": {
-    "DXswapLibrary": {
-      "address": "0xb73E00D36a43099d996a2E9B567dD91EFe3f53d9",
-      "constructorCode": "60566023600b82828239805160001a607314601657fe5b30600052607381538281f3fe",
-      "bodyBytecodeHash": "ee73aad5bd038ce2475e754f20e3ab26185542921d592fb0d073842a183ee677",
-      "localBytecodeHash": "4bffc79c339357805a6807e3a9bf61258bf4f865ed87546869cc5a2d72ac9c19",
-      "deployedBytecodeHash": "4bffc79c339357805a6807e3a9bf61258bf4f865ed87546869cc5a2d72ac9c19",
-      "types": {},
-      "storage": [],
-      "warnings": {
-        "hasConstructor": false,
-        "hasSelfDestruct": false,
-        "hasDelegateCall": false,
-        "hasInitialValuesInDeclarations": false,
-        "uninitializedBaseContracts": []
-      }
-    },
-    "DXswapOracleLibrary": {
-      "address": "0x9Fd0A02a3d67F5C1ABa3f713D28aBc1baBf74145",
-      "constructorCode": "60566023600b82828239805160001a607314601657fe5b30600052607381538281f3fe",
-      "bodyBytecodeHash": "ee73aad5bd038ce2475e754f20e3ab26185542921d592fb0d073842a183ee677",
-      "localBytecodeHash": "4bffc79c339357805a6807e3a9bf61258bf4f865ed87546869cc5a2d72ac9c19",
-      "deployedBytecodeHash": "4bffc79c339357805a6807e3a9bf61258bf4f865ed87546869cc5a2d72ac9c19",
-      "types": {},
-      "storage": [],
-      "warnings": {
-        "hasConstructor": false,
-        "hasSelfDestruct": false,
-        "hasDelegateCall": false,
-        "hasInitialValuesInDeclarations": false,
-        "uninitializedBaseContracts": []
-      }
-    }
-  },
-  "solidityLibs": {},
-  "proxies": {
-    "dxswap-periphery/DXswapRouter": [
-      {
-        "address": "0xf8Ecb734b74209E651E1A67Df0E9F9866E51B169",
-        "kind": "NonProxy",
-        "bytecodeHash": "dbdf8f09e50181b54b655f7890d7d43f3594c1b8b3527b55b413063677b205fd"
-      }
-    ]
-  },
-  "manifestVersion": "2.2",
-  "version": "0.2.0"
-}
