+++ package.json
@@ -14,10 +14,9 @@
     "contracts"
   ],
   "dependencies": {
-    "dxswap-core": "git://github.com/levelkdev/dxswap-core.git#v0.1.1"
+    "dxswap-core": "git://github.com/levelkdev/dxswap-core.git#3598b8b97fed373bb69db3344b7853adeefb0f9a"
   },
   "devDependencies": {
-    "@openzeppelin/cli": "^2.8.2",
     "@types/chai": "^4.2.6",
     "@types/mocha": "^5.2.7",
     "chai": "^4.2.0",
@@ -40,7 +39,7 @@
     "lint:fix": "yarn prettier ./test/*.ts --write",
     "clean": "rimraf ./build/",
     "precompile": "yarn clean",
-    "compile": "npx oz compile",
+    "compile": "waffle .waffle.json",
     "pretest": "yarn compile",
     "test": "mocha",
     "prepublishOnly": "yarn test",
