+++ test/DXswapRouter.spec.ts
@@ -8,8 +8,8 @@ import { ecsign } from 'ethereumjs-util'
 import { expandTo18Decimals, getApprovalDigest, mineBlock, MINIMUM_LIQUIDITY } from './shared/utilities'
 import { dxswapFixture } from './shared/fixtures'
 
-import IDXswapPair from 'dxswap-core/build/contracts/IDXswapPair.json'
-import DeflatingERC20 from '../build/contracts/DeflatingERC20.json'
+import IDXswapPair from 'dxswap-core/build/IDXswapPair.json'
+import DeflatingERC20 from '../build/DeflatingERC20.json'
 
 chai.use(solidity)
